refactory params

+++ electron/ipc/db/db.ts
@@ -14,39 +14,37 @@ import {
   IGetFilesRequestData,
 } from "../../../src/shared/interface/ipc/db";
 import { IDirectoryEntity } from "../../../src/shared/interface/db/directory";
+import { fetchDatabase } from "../../src/db";
 
 export const databaseChannel: IIpcChannel = {
   getName: () => IpcChannel.DATABASE,
   async handle(event: IpcMainEvent, request: IIpcRequest) {
-    switch (request.action) {
-      case DatabaseAction.GET_ROOT_DIRECTORY: {
-        const response = await DirectoryProvider.getRootDirectory();
-        // root should be an element
-        event.sender.send(request.responseChannel as string, response);
-        break;
-      }
-      case DatabaseAction.GET_FILES: {
-        const response = await FileProvider.getFiles(
-          (request.data as IGetFilesRequestData).directoryId
+    if (!request.responseChannel) {
+      request.responseChannel = `${this.getName()}_response`;
+    }
+    if (request.action === DatabaseAction.FETCH_DATABASE) {
+      await fetchDatabase();
+    } else if (request.action === DatabaseAction.GET_ROOT_DIRECTORY) {
+      const response = await DirectoryProvider.getRootDirectory();
+      // root should be an element
+      event.sender.send(request.responseChannel as string, response);
+    } else if (request.action === DatabaseAction.GET_FILES) {
+      const response = await FileProvider.getFiles(
+        (request.data as IGetFilesRequestData).directoryId
+      );
+      event.sender.send(request.responseChannel as string, response);
+    } else if (request.action === DatabaseAction.DELETE_DIRECTORY) {
+      const response = await DirectoryProvider.deleteDirectory(
+        (request.data as IDeleteDirectoryRequestData).id
+      );
+      event.sender.send(request.responseChannel as string, response);
+    } else if (request.action === DatabaseAction.CREATE_DIRECTORY) {
+      const response: IDirectoryEntity =
+        await DirectoryProvider.createDirectory(
+          (request.data as ICreateDirectoryRequestData).parentId,
+          (request.data as ICreateDirectoryRequestData).name
         );
-        event.sender.send(request.responseChannel as string, response);
-        break;
-      }
-      case DatabaseAction.DELETE_DIRECTORY: {
-        const response = await DirectoryProvider.deleteDirectory(
-          (request.data as IDeleteDirectoryRequestData).id
-        );
-        event.sender.send(request.responseChannel as string, response);
-        break;
-      }
-      case DatabaseAction.CREATE_DIRECTORY: {
-        const response: IDirectoryEntity =
-          await DirectoryProvider.createDirectory(
-            (request.data as ICreateDirectoryRequestData).parentId,
-            (request.data as ICreateDirectoryRequestData).name
-          );
-        event.sender.send(request.responseChannel as string, response);
-      }
+      event.sender.send(request.responseChannel as string, response);
     }
   },
 };
